Extract Tree Drag and Drop methods to own file
